refactor(models): move AttackPlan validation + finalisation shapes to models/attack.py
AttackPlanFinalisationError, AttackPlanValidationIssue, and
AttackPlanValidationReport were defined in tools/research_tools.py - same
shape of smell as AttackPlan itself living next to client code. Move all
three to models/attack.py where the rest of the AttackPlan domain shapes
already live, and have tools/research_tools.py import them. The validator
and finaliser functions stay where they are - they are behaviour, not data.
