+++ squad/vulnerability_researcher/__init__.py
@@ -10,17 +10,18 @@ from typing import Protocol, cast
 from crewai.tools import tool
 
 from models import Severity
-from models.attack import AttackPlan, AttackPlanItem
+from models.attack import (
+    AttackPlan,
+    AttackPlanFinalisationError,
+    AttackPlanItem,
+)
 from models.h1 import Programme
 from squad import SquadMember, read_run_file_tool, read_run_filelist_tool
 from tools import cwe_data, http, owasp_data
 from tools.h1_api import h1
 from tools.pentest import lookup_cve
 from tools.report_tools import calculate_cvss_score
-from tools.research_tools import (
-    AttackPlanFinalisationError,
-    finalise_research,
-)
+from tools.research_tools import finalise_research
 from tools.research_vocab import compose_research_brief_doc
 from tools.triage_tools import (
     DiscardEntry,
